ADD: 회원 가입을 위해 firebase 연동
