+++ .gitignore
@@ -29,3 +29,4 @@ Thumbs.db
 
 /app/android-key.keystore
 gradle.properties
+app/google-services.json
