+++ app/src/main/java/net/ifmain/monologue/data/DiaryEntryDto.kt
@@ -1,0 +1,8 @@
+package net.ifmain.monologue.data
+
+data class DiaryEntryDto(
+    val date: String,
+    val text: String,
+    val mood: String?,
+    val userId: String
+)
