+++ app/src/main/java/net/ifmain/monologue/data/UserEntryDto.kt
@@ -1,0 +1,7 @@
+package net.ifmain.monologue.data
+
+data class UserEntryDto(
+    val name: String,
+    val email: String,
+    val password: String
+)
