+++ gradle/libs.versions.toml
@@ -1,5 +1,6 @@
 [versions]
 agp = "8.9.1"
+firebaseBom = "33.12.0"
 kotlin = "2.0.21"
 coreKtx = "1.15.0"
 converterGson = "2.11.0"
@@ -14,6 +15,8 @@ retrofit = "2.11.0"
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
+firebase-analytics = { module = "com.google.firebase:firebase-analytics" }
+firebase-bom = { module = "com.google.firebase:firebase-bom", version.ref = "firebaseBom" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
 androidx-junit = { group = "androidx.test.ext", name = "junit", version.ref = "junitVersion" }
 androidx-espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "espressoCore" }
